PIM-3133: Move files from catalog to enrich bundle
