+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -25,4 +25,3 @@ parameters:
     pim_catalog.entity.metric.class:                 Pim\Bundle\CatalogBundle\Model\Metric
     pim_catalog.entity.product_media.class:          Pim\Bundle\CatalogBundle\Model\ProductMedia
     pim_catalog.entity.product_price.class:          Pim\Bundle\CatalogBundle\Model\ProductPrice
-    pim_catalog.entity.sequential_edit.class:        Pim\Bundle\CatalogBundle\Entity\SequentialEdit
