+++ src/Pim/Bundle/CatalogBundle/Resources/config/factories.yml
@@ -5,7 +5,6 @@ parameters:
     pim_catalog.factory.attribute_type.class:        Pim\Bundle\CatalogBundle\AttributeType\AttributeTypeFactory
     pim_catalog.factory.media.class:                 Pim\Bundle\CatalogBundle\Factory\MediaFactory
     pim_catalog.factory.metric.class:                Pim\Bundle\CatalogBundle\Factory\MetricFactory
-    pim_catalog.factory.sequential_edit.class:       Pim\Bundle\CatalogBundle\Factory\SequentialEditFactory
 
 services:
     pim_catalog.factory.family:
@@ -36,8 +35,3 @@ services:
         class: %pim_catalog.factory.metric.class%
         arguments:
             - %pim_catalog.entity.metric.class%
-
-    pim_catalog.factory.sequential_edit:
-        class: %pim_catalog.factory.sequential_edit.class%
-        arguments:
-            - %pim_catalog.entity.sequential_edit.class%
