+++ src/Pim/Bundle/CatalogBundle/Resources/config/managers.yml
@@ -14,7 +14,6 @@ parameters:
     pim_catalog.manager.product.class:             Pim\Bundle\CatalogBundle\Manager\ProductManager
     pim_catalog.manager.product_mass_action.class: Pim\Bundle\CatalogBundle\Manager\ProductMassActionManager
     pim_catalog.manager.product_category.class:    Pim\Bundle\CatalogBundle\Manager\ProductCategoryManager
-    pim_catalog.manager.sequential_edit.class:     Pim\Bundle\CatalogBundle\Manager\SequentialEditManager
 
 services:
     # Managers
@@ -127,10 +126,3 @@ services:
             - '@pim_catalog.repository.association_type'
             - '@doctrine.orm.entity_manager'
             - '@event_dispatcher'
-
-    pim_catalog.manager.sequential_edit:
-        class: %pim_catalog.manager.sequential_edit.class%
-        arguments:
-            - '@doctrine.orm.entity_manager'
-            - '@pim_catalog.repository.sequential_edit'
-            - '@pim_catalog.factory.sequential_edit'
