+++ src/Pim/Bundle/CatalogBundle/Resources/config/repositories.yml
@@ -11,7 +11,6 @@ parameters:
     pim_catalog.repository.group_type.class:       Pim\Bundle\CatalogBundle\Entity\Repository\GroupTypeRepository
     pim_catalog.repository.locale.class:           Pim\Bundle\CatalogBundle\Entity\Repository\LocaleRepository
     pim_catalog.repository.product_value.class:    Doctrine\ORM\EntityRepository
-    pim_catalog.repository.sequential_edit.class:  Pim\Bundle\CatalogBundle\Entity\Repository\SequentialEditRepository
 
 services:
     pim_catalog.repository.association_type:
@@ -119,11 +118,3 @@ services:
     pim_catalog.doctrine.product_query_builder:
         class: %pim_catalog.doctrine.product_query_builder.class%
         arguments: ['@pim_catalog.context.catalog']
-
-    pim_catalog.repository.sequential_edit:
-        class: %pim_catalog.repository.sequential_edit.class%
-        factory_service: doctrine.orm.entity_manager
-        factory_method: getRepository
-        arguments: [%pim_catalog.entity.sequential_edit.class%]
-        tags:
-            - { name: 'pim_repository' }
