+++ src/Pim/Bundle/EnrichBundle/DependencyInjection/PimEnrichExtension.php
@@ -38,6 +38,9 @@ class PimEnrichExtension extends Extension
         $loader->load('event_listeners.yml');
         $loader->load('form_subscribers.yml');
         $loader->load('resolvers.yml');
+        $loader->load('managers.yml');
+        $loader->load('entities.yml');
+        $loader->load('repositories.yml');
 
         if ($config['record_mails']) {
             $loader->load('mail_recorder.yml');
