+++ src/Pim/Bundle/EnrichBundle/Resources/config/controllers.yml
@@ -187,5 +187,5 @@ services:
             - '@oro_datagrid.mass_action.parameters_parser'
             - '@pim_datagrid.extension.mass_action.dispatcher'
             - %mass_edit_limit%
-            - '@pim_catalog.manager.sequential_edit'
+            - '@pim_enrich.manager.sequential_edit'
             - '@pim_user.context.user'
