+++ src/Pim/Bundle/EnrichBundle/Resources/config/doctrine/SequentialEdit.orm.yml
@@ -1,7 +1,7 @@
-Pim\Bundle\CatalogBundle\Entity\SequentialEdit:
+Pim\Bundle\EnrichBundle\Entity\SequentialEdit:
     type: entity
-    table: pim_catalog_sequential_edit
-    repositoryClass: Pim\Bundle\CatalogBundle\Entity\Repository\SequentialEditRepository
+    table: pim_enrich_sequential_edit
+    repositoryClass: Pim\Bundle\EnrichBundle\Entity\Repository\SequentialEditRepository
     fields:
         id:
             type: integer
@@ -10,6 +10,7 @@ Pim\Bundle\CatalogBundle\Entity\SequentialEdit:
                 strategy: AUTO
         productSet:
             type: array
+            column: product_set
     oneToOne:
         user:
             targetEntity: Symfony\Component\Security\Core\User\UserInterface
