+++ src/Pim/Bundle/EnrichBundle/Resources/config/factories.yml
@@ -1,5 +1,6 @@
 parameters:
     pim_enrich.form.factory.product_value_form_factory: Pim\Bundle\EnrichBundle\Form\Factory\ProductValueFormFactory
+    pim_enrich.factory.sequential_edit.class:           Pim\Bundle\EnrichBundle\Factory\SequentialEditFactory
 
 services:
     pim_enrich.factory.product_value_form_factory:
@@ -8,3 +9,8 @@ services:
             - '@form.factory'
             - '@pim_catalog.factory.attribute_type'
             - '@event_dispatcher'
+
+    pim_enrich.factory.sequential_edit:
+        class: %pim_enrich.factory.sequential_edit.class%
+        arguments:
+            - %pim_enrich.entity.sequential_edit.class%
