+++ src/Pim/Bundle/EnrichBundle/Resources/config/managers.yml
@@ -1,0 +1,10 @@
+parameters:
+    pim_enrich.manager.sequential_edit.class: Pim\Bundle\EnrichBundle\Manager\SequentialEditManager
+
+services:
+    pim_enrich.manager.sequential_edit:
+        class: %pim_enrich.manager.sequential_edit.class%
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '@pim_enrich.repository.sequential_edit'
+            - '@pim_enrich.factory.sequential_edit'
