+++ src/Pim/Bundle/EnrichBundle/Resources/config/repositories.yml
@@ -1,0 +1,11 @@
+parameters:
+    pim_enrich.repository.sequential_edit.class: Pim\Bundle\EnrichBundle\Entity\Repository\SequentialEditRepository
+
+services:
+    pim_enrich.repository.sequential_edit:
+        class: %pim_enrich.repository.sequential_edit.class%
+        factory_service: doctrine.orm.entity_manager
+        factory_method: getRepository
+        arguments: [%pim_enrich.entity.sequential_edit.class%]
+        tags:
+            - { name: 'pim_repository' }
